+++ source/geh_common/src/geh_common/domain/types/__init__.py
@@ -1,4 +1,5 @@
 from .asset_type import AssetType
+from .business_reason import BusinessReason
 from .charge_quality import ChargeQuality
 from .charge_resolution import ChargeResolution
 from .charge_type import ChargeType
@@ -14,12 +15,14 @@ from .net_settlement_group import NetSettlementGroup
 from .orchestration_type import OrchestrationType
 from .product import Product
 from .quantity_quality import QuantityQuality
+from .relation_type import RelationType
 from .settlement_group import SettlementGroup
 from .settlement_method import SettlementMethod
 from .time_resolution import TimeResolution
 
 __all__ = [
     "AssetType",
+    "BusinessReason",
     "ConnectionState",
     "ConnectionType",
     "DisconnectionType",
@@ -29,6 +32,7 @@ __all__ = [
     "OrchestrationType",
     "Product",
     "QuantityQuality",
+    "RelationType",
     "EnergyUnit",
     "SettlementMethod",
     "MeteringPointResolutionLegacy",